评论

+++ mall-portal/src/main/java/com/macro/mall/pay/service/H5PayService.java
@@ -3,5 +3,5 @@ package com.macro.mall.pay.service;
 import java.util.Map;
 
 public interface H5PayService {
-    Map<String, String> getPayParam(int amount, String orderId);
+    Map<String, String> getPayParam(double amount, String orderId);
 }
